3lab/requirements.txt
alembic==1.15.2
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
billiard==4.2.1
celery==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.12
greenlet==3.1.1
h11==0.14.0
idna==3.10
kombu==5.5.3
Mako==1.3.9
MarkupSafe==3.0.2
passlib==1.7.4
prompt_toolkit==3.0.51
psutil==7.0.0
psycopg2-binary==2.9.10
pyasn1==0.4.8
pydantic==2.11.1
pydantic-settings==2.8.1
pydantic_core==2.33.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.20
redis==6.2.0
redislite==6.2.912183
requests==2.32.3
rsa==4.9
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.1
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.0
vine==5.1.0
wcwidth==0.2.13
websockets==10.4
